--- pkg/s3wrapper/errors.go
@@ -1,9 +1,0 @@
-package s3wrapper
-
-import "fmt"
-
-type NotFound string
-
-func (f NotFound) Error() string {
-	return fmt.Sprintf("object %s was not found", string(f))
-}
